# Filter gaps before colormap autoscale