--- a/dioptas/widgets/plot_widgets/ColormapPopup.py
+++ b/dioptas/widgets/plot_widgets/ColormapPopup.py
@@ -128,6 +128,12 @@ class ColormapPopup(QtWidgets.QFrame):
 
         self._resetButtonGroup.buttonClicked.connect(self._autoscaleRequested)
 
+        self._filterGapsCheckBox = QtWidgets.QCheckBox(self)
+        self._filterGapsCheckBox.setToolTip("Toggle detector gaps value filtering")
+        self._filterGapsCheckBox.setChecked(utils.auto_level.filter_dummy)
+        self._filterGapsCheckBox.toggled.connect(self._autoscaleRequested)
+        layout.addRow('Filter gaps:', self._filterGapsCheckBox)
+
         buttonBox = QtWidgets.QDialogButtonBox(parent=self)
         buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Close)
         closeButton = buttonBox.button(QtWidgets.QDialogButtonBox.Close)
@@ -225,6 +231,7 @@ class ColormapPopup(QtWidgets.QFrame):
 
     def _autoscaleRequested(self, *args):
         utils.auto_level.mode = self._getResetMode()
+        utils.auto_level.filter_dummy = self._filterGapsCheckBox.isChecked()
         colormapRange = utils.auto_level.get_range(self.getData(copy=False))
         if colormapRange is None:
              return