This PR ~builds on PR #166 (and includes) and~ adds probing and filtering of gap values before performing the colormap autoscale.

~The 3 added commits are: https://github.com/Dioptas/Dioptas/commit/8df3ad6cb5d6e739e5a4091ae8e8feac73096a5e https://github.com/Dioptas/Dioptas/commit/127b924653ed8c6564810801520aedda464f5373 https://github.com/Dioptas/Dioptas/commit/0263dda7284be6be3344bc4ee78625c47a3d0151~

I tried to retrieve the selected `Detector` in the calibration tab and use pyFAI `Detector.dynamic_mask` but there's a few issues: retrieving the current `Detector` from the `AutoLevel`; the image data looks to be converted to `float32` for display which prevent `Detector.dynamic_mask` to work correctly; the image is transposed for compatibility with pyqtgraph and there is possible transformations to take care of.
So finally, probing the gap value is much simpler. The limitation is that it does not take care of other dummy pixel values.