@CPrescher 2023-10-30
Hi, functionality looks good. Did you try to measure the speed of the algorithm? Would it be really bad to just have it enabled all the time as default?