@t20100 2023-11-06
A quick benchmark:
```python
import fabio
from dioptas.widgets.plot_widgets.utils import detector_gap_mask, auto_level

data = fabio.open("dioptas/tests/data/CeO2_Pilatus1M.tif").data
```

```
%timeit detector_gap_mask(data)
599 µs ± 2.71 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

And for the overall autoscale time:
* with filter:
  ```
  auto_level.filter_dummy = True
  %timeit auto_level.get_range(data)
  14.7 ms ± 16.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
  ```
* without:
  ```
  auto_level.filter_dummy = False
  %timeit auto_level.get_range(data)
  14.1 ms ± 49 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
 ```

I was a bit concerned with wrong detection that might screw-up the autoscale since the detection is very basic. But that can be made the default and in case of issue to be disabled manually.